Flask-Server/requirements.txt
bcrypt==3.1.3
cryptography
Flask==0.12
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.1
marshmallow==2.2.1
marshmallow-jsonapi==0.3.0
pytest==3.1.2
pytest-cov==2.5.1
SQLAlchemy==1.1.6
Werkzeug==0.11.15
PyJWT==1.5.3
